backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
boto3==1.35.88
botocore==1.35.88
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
requests>=2.31.0
